pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astraut-risk-reasoner"
version = "0.1.0"
description = "Astraut Risk Reasoner CLI for practical SME cybersecurity risk assessments"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Astraut Solutions" }
]
keywords = ["cybersecurity", "risk", "sme", "cli", "zero-trust", "llm"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "groq>=0.9.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "black>=24.8.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/astraut-solutions/astraut-risk-reasoner"
Repository = "https://github.com/astraut-solutions/astraut-risk-reasoner"
Issues = "https://github.com/astraut-solutions/astraut-risk-reasoner/issues"
Changelog = "https://github.com/astraut-solutions/astraut-risk-reasoner/releases"

[project.scripts]
astraut-risk = "astraut_risk.__main__:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
astraut_risk = ["frameworks/*.yaml"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
